Update component index to name for readability
