+++ client/src/components/ProgressBar.js
@@ -1,0 +1,8 @@
+import React from "react";
+import { Progress } from "antd";
+
+const ProgressBar = () => {
+    return <Progress percent={50} status="active" />;
+};
+
+export default ProgressBar;
