+++ client/src/pages/Quiz.js
@@ -1,0 +1,12 @@
+import React from "react";
+import ProgressBar from "../components/ProgressBar";
+
+const Quiz = () => {
+    return (
+        <div className="page">
+            <ProgressBar />
+        </div>
+    );
+};
+
+export default Quiz;
